Rename mainnavigator folder to camelCase
