+++ src/components/mainNavigator/MainNavigator.tsx
