+++ src/hooks/usePasswordMatch.ts
@@ -1,5 +1,5 @@
 import { useState } from "react";
-import { FormData } from "../Types/types";
+import { FormData } from "../types/types";
 
 export const usePasswordMatch = () => {
     const [isPasswordMatch, setIsPasswordMatch] = useState(true);
